+++ test/tests.ts
@@ -40,3 +40,4 @@ import './mockNetworkInterface';
 import './graphqlSubscriptions';
 import './batchedNetworkInterface';
 import './ObservableQuery';
+import './subscribeToMore';